Merged PR 14280: Demonstration dataset preprocessing
This PR contains the notebook and the class that provide the implementation of the data preprocessing. The demonstrations are aligned in time though the dynamic time warping algorithm. Then they are sampled with a desired sampling frequency and padding is used to bring them on the same length.

Related work items: #109693