Generalized demonstration_player.py to take Trajectories object as input, no need to distinguish between reading from file and reading pipeline output
This commit restructures the DemonstrationPlayer class. Now a Trajectories object is passed as input and so it is possible to reproduce both trajectory
demonstrations and trajectories generated by the dynamic movement primitives method. Moreover, the error between consecutive joint angles vector is
computed as the norm of the vecotr instead of using the Lie algebra.

Adds #110744