Align trajectory on static reference
Previously, trajectories were aligned onto the reference trajectory
and the reference was then aligned onto the last trajectory in the list.
This behavior meant that the other of the trajectory had an impact
on the GMM that would be extract later since the reference would
be aligned differentlz depending on the order.
Since this commit, trajectories are aligned to the reference and
the reference is not aligned to anything anymore. Thus, unit tests
are fully reproducible regardless of the order of the trajectories.